saving before changing CategoryName to CategoryId

+++ Business/Services/CategoryService.cs
@@ -9,10 +9,12 @@ namespace Business.Services;
 public class CategoryService
 {
     private readonly ICategoryRepository _categoryRepository;
+    private readonly UserService _userService;
 
-    public CategoryService(ICategoryRepository categoryRepository)
+    public CategoryService(ICategoryRepository categoryRepository, UserService userService)
     {
         _categoryRepository = categoryRepository;
+        _userService = userService;
     }
 
     public CategoryEntity GetOrCreateCategory(ProductRegistrationDto product)
@@ -36,12 +38,13 @@ public class CategoryService
         return null!;
     }
 
-    public CategoryDto GetOrCreateCategory(CategoryDto category)
+    public CategoryDto CreateCategory(CategoryDto category)
     {
         try
         {
+            var activeUser = _userService.FindRoleOfActiveUser();
             var exisitingCategoryName = _categoryRepository.GetOne(x => x.CategoryName == category.CategoryName);
-            if (exisitingCategoryName == null)
+            if (exisitingCategoryName == null && activeUser.UserRoleName == "Admin")
             {
                 var newCategoryName = _categoryRepository.Create(new CategoryEntity()
                 {
@@ -49,21 +52,34 @@ public class CategoryService
                 });
                 return CategoryFactory.Create(newCategoryName);
             }
-            else
-                return CategoryFactory.Create(exisitingCategoryName);
         }
         catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
 
         return null!;
     }
-    
+
+    public CategoryDto GetCategory(CategoryDto category)
+    {
+        try
+        {
+            var exisitingCategory = _categoryRepository.GetOne(x => x.Id == category.Id);
+            if (exisitingCategory != null)
+            {
+                return CategoryFactory.Create(exisitingCategory);
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+
+        return null!;
+    }
+
     public IEnumerable<CategoryDto> GetAllCategories()
     {
         try
         {
-            var result = _categoryRepository.GetAll();
-            if (result != null)
-                return CategoryFactory.Create(result);
+            var exisitingCategories = _categoryRepository.GetAll();
+            if (exisitingCategories != null)
+                return CategoryFactory.Create(exisitingCategories);
         }
         catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
 
@@ -74,13 +90,14 @@ public class CategoryService
     {
         try
         {
-            var existingCategory = _categoryRepository.GetOne(x => x.CategoryName == existingDtoName.CategoryName);
-            if (existingCategory != null)
+            var activeUser = _userService.FindRoleOfActiveUser();
+            var existingCategory = _categoryRepository.GetOne(x => x.Id == existingDtoName.Id);
+            if (existingCategory != null && activeUser.UserRoleName == "Admin")
             {
                 CategoryEntity updatedEntity = new()
                 {
                     Id = existingCategory.Id,
-                    CategoryName = !string.IsNullOrEmpty(updatedDtoName.CategoryName) ? updatedDtoName.CategoryName : existingDtoName.CategoryName,
+                    CategoryName = !string.IsNullOrEmpty(updatedDtoName.CategoryName) ? updatedDtoName.CategoryName : existingCategory.CategoryName,
                 };
 
                 var result = _categoryRepository.Update(existingCategory, updatedEntity);
@@ -99,8 +116,9 @@ public class CategoryService
     {
         try
         {
+            var activeUser = _userService.FindRoleOfActiveUser();
             var existingCategory = _categoryRepository.GetOne(x => x.CategoryName == category.CategoryName && x.Id == category.Id);
-            if (existingCategory != null)
+            if (existingCategory != null && activeUser.UserRoleName == "Admin")
             {
                 var result = _categoryRepository.Delete(existingCategory);
                 if (result)
